uses double instead of floats

+++ include/Subsample_and_aggregate.hpp
@@ -13,24 +13,24 @@
 
 class Subsample_and_aggregate{
 public:
-    Subsample_and_aggregate(std::vector<float> (*to_compute)(std::vector<float>), std::vector<float> (*Aggregate)(std::vector<std::vector<float>>, float delta_agg, float epsilon), std::vector<float> db, int number_of_partition, float epsilon, float delta, float delta_agg);
+    Subsample_and_aggregate(std::vector<double> (*to_compute)(std::vector<double>), std::vector<double> (*Aggregate)(std::vector<std::vector<double>>, double delta_agg, double epsilon), std::vector<double> db, int number_of_partition, double epsilon, double delta, double delta_agg);
     
-    std::vector<float> get_result();
+    std::vector<double> get_result();
 private:
     int number_of_partition;
-    std::vector<float> db;
-    std::vector<std::vector<float>> randomly_partition();
-    float epsilon;
-    float delta;
-    std::vector<float> (*to_compute)(std::vector<float>);
-    std::vector<float> (*Aggregate)(std::vector<std::vector<float>>, float delta_agg, float epsilon);
-    float delta_agg;
+    std::vector<double> db;
+    std::vector<std::vector<double>> randomly_partition();
+    double epsilon;
+    double delta;
+    std::vector<double> (*to_compute)(std::vector<double>);
+    std::vector<double> (*Aggregate)(std::vector<std::vector<double>>, double delta_agg, double epsilon);
+    double delta_agg;
 };
 
-std::vector<float> deciles(std::vector<float> db);
-float mean_10(std::vector<float> data);
-std::vector<float> Aggregate_10_mean(std::vector<std::vector<float>> samples_computed);
-void test_deciles_subsample_and_aggregate(int n, int k, float epsilon, float delta);
-float square_mean_error(std::vector<float> t1, std::vector<float> t2);
+std::vector<double> deciles(std::vector<double> db);
+double mean_10(std::vector<double> data);
+std::vector<double> Aggregate_10_mean(std::vector<std::vector<double>> samples_computed);
+void test_deciles_subsample_and_aggregate(int n, int k, double epsilon, double delta);
+double square_mean_error(std::vector<double> t1, std::vector<double> t2);
 
 #endif /* Subsample_and_aggregate_hpp */
